// frontendclient/src/app/api/generated/experiences/experiences.service.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import {
  HttpClient
} from '@angular/common/http';

import {
  Injectable,
  inject
} from '@angular/core';

import type {
  Experience
} from '../../models';

import { customHttpClient } from '../../custom-http-client';






@Injectable({ providedIn: 'root' })
export class ExperiencesService {
  private readonly http = inject(HttpClient);
/**
 * Retrieve a specific work experience by its ID
 * @summary Get experience by ID
 */
 getExperienceById<TData = Experience>(
    id: string,
 ) {
      return customHttpClient<TData>(
      {url: `/api/experiences/${id}`, method: 'GET'
    },
      this.http,
      );
    }
  /**
 * Update an existing work experience
 * @summary Update experience
 */
 updateExperience<TData = Experience>(
    id: string,
    experience: Experience,
 ) {
      return customHttpClient<TData>(
      {url: `/api/experiences/${id}`, method: 'PUT',
      headers: {'Content-Type': 'application/json', },
      data: experience
    },
      this.http,
      );
    }
  /**
 * Remove a work experience from the portfolio
 * @summary Delete experience
 */
 deleteExperience<TData = void>(
    id: string,
 ) {
      return customHttpClient<TData>(
      {url: `/api/experiences/${id}`, method: 'DELETE'
    },
      this.http,
      );
    }
  /**
 * Retrieve a list of all work experiences
 * @summary Get all experiences
 */
 getAllExperiences<TData = Experience[]>(
    
 ) {
      return customHttpClient<TData>(
      {url: `/api/experiences`, method: 'GET'
    },
      this.http,
      );
    }
  /**
 * Add a new work experience
 * @summary Create new experience
 */
 createExperience<TData = Experience>(
    experience: Experience,
 ) {
      return customHttpClient<TData>(
      {url: `/api/experiences`, method: 'POST',
      headers: {'Content-Type': 'application/json', },
      data: experience
    },
      this.http,
      );
    }
  };

export type GetExperienceByIdClientResult = NonNullable<Experience>
export type UpdateExperienceClientResult = NonNullable<Experience>
export type DeleteExperienceClientResult = NonNullable<void>
export type GetAllExperiencesClientResult = NonNullable<Experience[]>
export type CreateExperienceClientResult = NonNullable<Experience>