// frontendclient/src/app/api/models/experience.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface Experience {
  id?: string;
  title?: string;
  company?: string;
  location?: string;
  startDate?: string;
  endDate?: string;
  description?: string;
  lastUpdated?: string;
  isVisible?: boolean;
  current?: boolean;
}